GetScores functionality works
