+++ Bancho/Core/Helpers/CanNotAccessBanchoException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace osuBancho.Helpers
+{
+    [Serializable]
+    class CanNotAccessBanchoException: Exception
+    {
+    }
+}
