+++ Bancho/Core/Helpers/Counting.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace osuBancho.Helpers
+{
+    class Counting
+    {
+        public static Int32 Count(string orig, string find)
+        {
+            var s2 = orig.Replace(find, "");
+            return (orig.Length - s2.Length) / find.Length;
+        }
+    }
+}
