+++ Bancho/Core/Helpers/ScoreHelper.cs
