--- Bancho/Helpers/CanNotAccessBanchoException.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace osuBancho.Helpers
-{
-    [Serializable]
-    class CanNotAccessBanchoException: Exception
-    {
-    }
-}
